requirements.txt
appdirs==1.4.3
Django==1.10.6
django-crispy-forms==1.6.1
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
six==1.10.0
